requirements-dev.txt
coverage==7.5.*
jsonschema==4.23.*
mypy==1.14.*
network
openapi-core==0.19.*
pre-commit==4.1.*
pyserial==3.5.*
pytest==8.3.*
pytest-cov==5.0.*
pyyaml==6.0.*
sphinx==7.4.*
sphinx-rtd-theme==2.0.*
types-requests==2.32.*
